requirements.txt
selenium>=4.15.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
webdriver-manager>=4.0.1
lxml>=4.9.3
requests>=2.31.0
